backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.23.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.7
alembic==1.11.1
pydantic==1.10.9
python-dotenv==1.0.0
celery==5.3.0
redis==4.6.0
requests==2.31.0
aiofiles==23.1.0
python-multipart==0.0.6
